+++ R/custom_microorganisms.R
@@ -247,19 +247,14 @@ add_custom_microorganisms <- function(x) {
     "CUSTOM",
     seq.int(from = current + 1, to = current + nrow(x), by = 1),
     "_",
-    toupper(unname(abbreviate(
-      gsub(
-        " +", " _ ",
-        gsub(
-          "[^A-Za-z0-9-]", " ",
-          trimws2(paste(x$genus, x$species, x$subspecies))
-        )
-      ),
-      minlength = 10
-    )))
-  )
+    trimws(
+      paste(abbreviate_mo(x$genus, 5),
+            abbreviate_mo(x$species, 4, hyphen_as_space = TRUE),
+            abbreviate_mo(x$subspecies, 4, hyphen_as_space = TRUE),
+            sep = "_"),
+      whitespace = "_"))
   stop_if(anyDuplicated(c(as.character(AMR_env$MO_lookup$mo), x$mo)), "MO codes must be unique and not match existing MO codes of the AMR package")
-
+  
   # add to package ----
   AMR_env$custom_mo_codes <- c(AMR_env$custom_mo_codes, x$mo)
   class(AMR_env$MO_lookup$mo) <- "character"
@@ -305,4 +300,27 @@ clear_custom_microorganisms <- function() {
   AMR_env$mo_previously_coerced <- AMR_env$mo_previously_coerced[which(AMR_env$mo_previously_coerced$mo %in% AMR_env$MO_lookup$mo), , drop = FALSE]
   AMR_env$mo_uncertainties <- AMR_env$mo_uncertainties[0, , drop = FALSE]
   message_("Cleared ", nr2char(n - n2), " custom record", ifelse(n - n2 > 1, "s", ""), " from the internal `microorganisms` data set.")
+}
+
+abbreviate_mo <- function(x, minlength = 5, prefix = "", hyphen_as_space = FALSE, ...) {
+  if (hyphen_as_space == TRUE) {
+    x <- gsub("-", " ", x, fixed = TRUE)
+  }
+  # keep a starting Latin ae
+  suppressWarnings(
+    gsub("(\u00C6|\u00E6)+",
+         "AE",
+         toupper(
+           paste0(prefix,
+                  abbreviate(
+                    gsub("^ae",
+                         "\u00E6\u00E6",
+                         x,
+                         ignore.case = TRUE),
+                    minlength = minlength,
+                    use.classes = TRUE,
+                    method = "both.sides",
+                    ...
+                  ))))
+  )
 }