+++ inst/tinytest/test-mo_property.R
@@ -98,10 +98,13 @@ expect_equal(names(mo_info("Escherichia coli")), c(
   "mo",
   "kingdom", "phylum", "class", "order",
   "family", "genus", "species", "subspecies",
-  "status", "synonyms", "gramstain", "url", "ref",
-  "snomed"
+  "status", "synonyms", "gramstain", "oxygen_tolerance",
+  "url", "ref", "snomed"
 ))
 expect_inherits(mo_info(c("Escherichia coli", "Staphylococcus aureus")), "list")
+
+expect_identical(mo_oxygen_tolerance(c("Klebsiella pneumoniae", "Clostridioides difficile")),
+                 c("aerobe", "anaerobe"))
 
 expect_equal(as.character(table(mo_pathogenicity(example_isolates$mo))),
              c("1561", "422", "1", "16"))